Refactor: Implement WorkOrderFormEnhanced refactor
This commit implements the refactoring plan for `WorkOrderFormEnhanced.tsx` by extracting logic into custom hooks (`useWorkOrderForm`, `useWorkOrderFormSubmission`, `useEquipmentSelection`) and creating specialized UI components for better modularity, testability, and maintainability.
