+++ src/components/work-orders/form/WorkOrderBasicFields.tsx
@@ -1,0 +1,103 @@
+import React from 'react';
+import { Card, CardContent } from "@/components/ui/card";
+import { Input } from "@/components/ui/input";
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
+import { Label } from "@/components/ui/label";
+import { WorkOrderFormData } from '@/hooks/useWorkOrderForm';
+
+interface WorkOrderBasicFieldsProps {
+  values: WorkOrderFormData;
+  errors: Record<string, string>;
+  setValue: (field: keyof WorkOrderFormData, value: any) => void;
+  preSelectedEquipment?: any;
+}
+
+export const WorkOrderBasicFields: React.FC<WorkOrderBasicFieldsProps> = ({
+  values,
+  errors,
+  setValue,
+  preSelectedEquipment
+}) => {
+  return (
+    <Card>
+      <CardContent className="pt-4 space-y-4">
+        <h3 className="font-medium text-sm text-muted-foreground uppercase tracking-wide">
+          Work Order Details
+        </h3>
+        
+        <div className="space-y-2">
+          <Label>Title *</Label>
+          <Input
+            placeholder={preSelectedEquipment ? 
+              `Maintenance for ${preSelectedEquipment.name}` : 
+              "Brief description of the work needed"
+            }
+            value={values.title || ''}
+            onChange={(e) => setValue('title', e.target.value)}
+          />
+          {errors.title && (
+            <p className="text-sm text-destructive">{errors.title}</p>
+          )}
+        </div>
+
+        <div className="space-y-2">
+          <Label>Priority *</Label>
+          <Select 
+            value={values.priority} 
+            onValueChange={(value) => setValue('priority', value as 'low' | 'medium' | 'high')}
+          >
+            <SelectTrigger>
+              <SelectValue placeholder="Select priority" />
+            </SelectTrigger>
+            <SelectContent>
+              <SelectItem value="low">
+                <div className="flex items-center gap-2">
+                  <div className="w-2 h-2 rounded-full bg-green-500"></div>
+                  Low Priority
+                </div>
+              </SelectItem>
+              <SelectItem value="medium">
+                <div className="flex items-center gap-2">
+                  <div className="w-2 h-2 rounded-full bg-yellow-500"></div>
+                  Medium Priority
+                </div>
+              </SelectItem>
+              <SelectItem value="high">
+                <div className="flex items-center gap-2">
+                  <div className="w-2 h-2 rounded-full bg-red-500"></div>
+                  High Priority
+                </div>
+              </SelectItem>
+            </SelectContent>
+          </Select>
+          {errors.priority && (
+            <p className="text-sm text-destructive">{errors.priority}</p>
+          )}
+        </div>
+
+        <div className="grid grid-cols-2 gap-4">
+          <div className="space-y-2">
+            <Label>Due Date</Label>
+            <Input
+              type="date"
+              value={values.dueDate || ''}
+              onChange={(e) => setValue('dueDate', e.target.value)}
+            />
+          </div>
+
+          <div className="space-y-2">
+            <Label>Estimated Hours</Label>
+            <Input
+              type="number"
+              min="0"
+              step="0.5"
+              placeholder="0"
+              value={values.estimatedHours || ''}
+              onChange={(e) => setValue('estimatedHours', parseFloat(e.target.value) || undefined)}
+            />
+          </div>
+        </div>
+      </CardContent>
+    </Card>
+  );
+};
